+++ src/main/java/com/skd/playeractivityview/mixin/client/NativeImageAccessorMixin.java
@@ -1,5 +1,8 @@
 package com.skd.playeractivityview.mixin.client;
 
+import com.mojang.blaze3d.platform.NativeImage;
+import org.spongepowered.asm.mixin.Mixin;
+
+@Mixin(NativeImage.class)
 public class NativeImageAccessorMixin {
-    public NativeImageAccessorMixin() {}
 }