+++ src/main/java/com/skd/playeractivityview/mixin/client/ScreenRenderWithTooltipMixin.java
@@ -1,5 +1,0 @@
-package com.skd.playeractivityview.mixin.client;
-
-public class ScreenRenderWithTooltipMixin {
-    public ScreenRenderWithTooltipMixin() {}
-}