+++ src/main/java/com/skd/playeractivityview/mixin/client/SetupRotationsInjectMixin.java
@@ -1,5 +1,0 @@
-package com.skd.playeractivityview.mixin.client;
-
-public class SetupRotationsInjectMixin {
-    public SetupRotationsInjectMixin() {}
-}